# Enqueue registered assets once