@oandregal 2023-04-28
Committed at https://core.trac.wordpress.org/changeset/55695