@oandregal 2023-05-24
Follow-up at https://github.com/WordPress/wordpress-develop/pull/4498